Time: 1 ms (99.85%), Space: 42.6 MB (52.17%) - LeetHub

+++ 3754-concatenate-non-zero-digits-and-multiply-by-sum-i/3754-concatenate-non-zero-digits-and-multiply-by-sum-i.java
@@ -14,6 +14,6 @@ class Solution {
             num/=10;
            
         }
-        return x*sum;
+        return (long)x*sum;
     }
 }
